Adds some vetting tools for event.
