+++ main.go
@@ -1,10 +1,7 @@
 package main
 
 import (
-	"encoding/json"
-	"io/ioutil"
 	"log"
-	"strings"
 )
 
 // TODO: whole twitter post as a field?
@@ -12,6 +9,7 @@ type event struct {
 	EventName     string `json:"EventName"`
 	VenueName     string `json:"VenueName"`
 	VenueLocation string `json:"VenueLocation"`
+	VenueDetails  string `json:"VenueDetails"`
 	StartTime     string `json:"StartTime"`
 	EndTime       string `json:"EndTime"`
 	Description   string `json:"Description"`
@@ -25,9 +23,10 @@ func main() {
 	setup()
 
 	// TODO: take in a flag with a file to post.
-	eventJSON, _ := readFileContents("sample-event.json")
-
-	eventEntry := readFromJSON(eventJSON)
+	eventEntry, err := fileToEvent("sample-event.json")
+	if err != nil {
+		log.Fatalln("Couldn't open file for reading.")
+	}
 
 	// These can be refactored into a single function that calls them all
 	sendToCalagator(eventEntry)
@@ -39,36 +38,6 @@ func main() {
 func setup() {
 	// envconfig bits for various integration, such as Meetups API
 	productionMode = false
-}
-
-func getDateFromDateTime(dateTime string) string {
-	date := strings.Split(dateTime, "|")[0]
-	return date
-}
-
-func getTimeFromDateTime(dateTime string) string {
-	time := strings.Split(dateTime, "|")[1]
-	return time
-}
-
-func readFileContents(file string) ([]byte, error) {
-	fileAsBytes, err := ioutil.ReadFile(file)
-
-	if err != nil {
-		log.Fatalln("Couldn't open ", file)
-	}
-
-	return fileAsBytes, nil
-}
-
-func readFromJSON(eventJSON []byte) event {
-	var eventEntry event
-	err := json.Unmarshal([]byte(eventJSON), &eventEntry)
-
-	if err != nil {
-		log.Fatal("Couldn't parse JSON file.")
-	}
-	return eventEntry
 }
 
 func sendToMeetup(eventEntry event) {
